refactor: remove all logic from program.cs
